requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
cbor==1.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
executing==2.2.0
filelock==3.17.0
filetype==1.2.0
FlagEmbedding @ git+https://github.com/FlagOpen/FlagEmbedding.git@fcdf889ec91edcd5278ba33a08c0665f4a59feb6
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
ijson==3.3.0
inscriptis==2.5.3
ipykernel==6.29.5
ipython==8.32.0
ir_datasets==0.5.9
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
llama-cloud==0.1.12
llama-cloud-services==0.6.0
llama-index==0.12.16
llama-index-agent-openai==0.4.5
llama-index-cli==0.4.0
llama-index-core==0.12.16.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.18
llama-index-multi-modal-llms-openai==0.4.3
llama-index-postprocessor-flag-embedding-reranker==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-milvus==0.5.0
llama-parse==0.6.0
lxml==5.3.0
lz4==4.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.11
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.25.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
openai==1.61.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pymilvus==2.5.4
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
streamlit==1.42.0
striprtf==0.0.26
sympy==1.13.1
tavily-python==0.5.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
trec-car-tools==2.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
unlzw3==0.2.3
urllib3==2.3.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zlib-state==0.1.9
